+++ lib/laflem/collections/python/virtualenv/__init__.py
@@ -5,3 +5,4 @@ from .cleanup import VirtualenvCleanupModule
 from .create import VirtualenvCreateModule
 from .list import VirtualenvListModule
 from .delete import VirtualenvDeleteModule
+from .activate import VirtualenvActivateModule