+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Features
 
 - [python][venv-create] Allow to specify a custom Python interpreter when creating virtual env with `--interpreter`
-
+- Add `venv-activate` module to `python` collection
 
 ## v0.4.0